Reformat the SupportedMixers configuration to match other configurations
