+++ packages/apps/src/configs/mixers/index.ts
@@ -1,0 +1,1 @@
+export * from './mixer-config';
