+++ packages/apps/src/initI18n.ts
@@ -1,9 +1,9 @@
+
 // @ts-nocheck
 // auto generate by buildI18n.js
 
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
-
 import resources from '../src/i18n/index.json';
 
 // for debug
@@ -15,5 +15,5 @@ i18n.use(initReactI18next).init({
   defaultNS: 'translations',
   fallbackLng: 'en',
   ns: ['apps', 'page-mixer', 'react-components'],
-  resources,
+  resources
 });
