+++ packages/react-environment/src/types/mixer-config.interface.ts
@@ -1,0 +1,10 @@
+export interface TornMixerEntry {
+  size: number;
+  address: string;
+  symbol: string;
+  createdAtBlock: number;
+}
+
+export interface MixerConfig {
+  tornMixers: TornMixerEntry[];
+}
