+++ packages/react-environment/src/webb-context/common.ts
@@ -2,6 +2,7 @@ import { AnchorConfigEntry } from '@webb-dapp/react-environment/types/anchor-con
 import { BridgeConfig } from '@webb-dapp/react-environment/types/bridge-config.interface';
 import { ChainConfig } from '@webb-dapp/react-environment/types/chain-config.interface';
 import { CurrencyConfig } from '@webb-dapp/react-environment/types/currency-config.interface';
+import { MixerConfig } from '@webb-dapp/react-environment/types/mixer-config.interface';
 import { WalletConfig } from '@webb-dapp/react-environment/types/wallet-config.interface';
 
 export type Chain = ChainConfig & {
@@ -15,4 +16,5 @@ export type AppConfig = {
   currencies: Record<number, CurrencyConfig>;
   bridgeByAsset: Record<number, BridgeConfig>;
   anchors: Record<number, AnchorConfigEntry[]>;
+  mixers: Record<number, MixerConfig>;
 };
